[GitWrapper] rename namespace
