+++ packages/GitWrapper/src/Event/GitOutputListenerInterface.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace GitWrapper\Event;
+namespace Symplify\GitWrapper\Event;
 
 /**
  * Interface implemented by output listeners.
