+++ packages/GitWrapper/src/Exception/GitException.php
@@ -1,0 +1,9 @@
+<?php declare(strict_types=1);
+
+namespace Symplify\GitWrapper\Exception;
+
+use RuntimeException;
+
+final class GitException extends RuntimeException
+{
+}
