+++ packages/GitWrapper/src/GitException.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace GitWrapper;
+namespace Symplify\GitWrapper;
 
 /**
  * Base class for exceptions thrown by the Git library.
