+++ packages/GitWrapper/src/GitProcess.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace GitWrapper;
+namespace Symplify\GitWrapper;
 
 use Symfony\Component\Process\Process;
