+++ packages/GitWrapper/tests/Event/TestBypassListener.php
@@ -1,8 +1,8 @@
 <?php declare(strict_types=1);
 
-namespace GitWrapper\Test\Event;
+namespace Symplify\GitWrapper\Test\Event;
 
-use GitWrapper\Event\GitEvent;
+use Symplify\GitWrapper\Event\GitEvent;
 
 final class TestBypassListener
 {
