+++ packages/GitWrapper/tests/Event/TestDispatcher.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace GitWrapper\Test\Event;
+namespace Symplify\GitWrapper\Test\Event;
 
 use Symfony\Component\EventDispatcher\EventDispatcher;
