+++ packages/GitWrapper/tests/Event/TestOutputListener.php
@@ -1,9 +1,9 @@
 <?php declare(strict_types=1);
 
-namespace GitWrapper\Test\Event;
+namespace Symplify\GitWrapper\Test\Event;
 
-use GitWrapper\Event\GitOutputEvent;
-use GitWrapper\Event\GitOutputListenerInterface;
+use Symplify\GitWrapper\Event\GitOutputEvent;
+use Symplify\GitWrapper\Event\GitOutputListenerInterface;
 
 final class TestOutputListener implements GitOutputListenerInterface
 {
@@ -12,7 +12,7 @@ final class TestOutputListener implements GitOutputListenerInterface
      */
     private $event;
 
-    
+
     public function getLastEvent(): GitWrapper\Event\GitOutputEvent
     {
         return $this->event;
